+++ src/Controller/Admin/PayPalConfigController.php
@@ -266,7 +266,15 @@ class PayPalConfigController extends AdminController
         if (!isset($conf['oscPayPalSetVaulting'])) {
             $conf['oscPayPalSetVaulting'] = false;
         }
-
+        if (!isset($conf['oscPayPalUseGooglePayAddress'])) {
+            $conf['oscPayPalUseGooglePayAddress'] = false;
+        }
+        if (!isset($conf['oscPayPalDefaultShippingPriceExpress'])) {
+            $conf['oscPayPalDefaultShippingPriceExpress'] = false;
+        } else {
+            $dAmount = (string) str_replace(',', '.', $conf['oscPayPalDefaultShippingPriceExpress']);
+            $conf['oscPayPalDefaultShippingPriceExpress'] = $dAmount;
+        }
         return $conf;
     }