+++ src/Core/Constants.php
@@ -12,7 +12,7 @@ namespace OxidSolutionCatalysts\PayPal\Core;
 class Constants
 {
     public const PAYPAL_JS_SDK_URL = 'https://www.paypal.com/sdk/js';
-    public const PAYPAL_INTEGRATION_DATE = '2020-07-29';
+    public const PAYPAL_INTEGRATION_DATE = '2024-04-26';
     public const PAYPAL_ORDER_INTENT_CAPTURE = 'CAPTURE';
     public const PAYPAL_ORDER_INTENT_AUTHORIZE = 'AUTHORIZE';
     public const SESSION_CHECKOUT_ORDER_ID = 'paypal-checkout-session';