+++ src/Core/OrderRequestFactory.php
@@ -101,6 +101,9 @@ class OrderRequestFactory
         $setVaulting = $moduleSettings->getIsVaultingActive();
         $selectedVaultPaymentSourceIndex = Registry::getSession()->getVariable("selectedVaultPaymentSourceIndex");
         $paymentId = Registry::getSession()->getVariable('paymentid');
+        if ($paymentId === PayPalDefinitions::GOOGLEPAY_PAYPAL_PAYMENT_ID) {
+            $request->payment_source = $this->getGooglePayPaymentSource($basket, 'google_pay');
+        }
         if ($paymentId === PayPalDefinitions::APPLEPAY_PAYPAL_PAYMENT_ID) {
             $request->payment_source = $this->getApplePayPaymentSource($basket,'apple_pay');
         }
@@ -197,6 +200,31 @@ class OrderRequestFactory
         return $paymentSource;
     }
 
+    protected function getGooglePayPaymentSource($basket, $requestName)
+    {
+        $user = $basket->getBasketUser();
+
+        $userName = $user->getFieldData('oxfname') . ' ' . $user->getFieldData('oxlname');
+
+        // get Billing CountryCode
+        $country = oxNew(Country::class);
+        $country->load($user->getFieldData('oxcountryid'));
+
+        // check possible deliveryCountry
+        $deliveryId = Registry::getSession()->getVariable("deladrid");
+        $deliveryAddress = oxNew(Address::class);
+        if ($deliveryId && $deliveryAddress->load($deliveryId)) {
+            $country->load($deliveryAddress->getFieldData('oxcountryid'));
+        }
+        $paymentSource = new \stdClass();
+
+// Dynamically adding properties to the stdClass object
+        $paymentSource->$requestName = new \stdClass();
+        $paymentSource->$requestName->attributes = new \stdClass();
+        $paymentSource->$requestName->attributes->verification = new \stdClass();
+        $paymentSource->$requestName->attributes->verification->method = 'SCA_ALWAYS';
+        return $paymentSource;
+    }
     /**
      * Sets application context
      *