+++ src/Model/Order.php
@@ -148,13 +148,14 @@ class Order extends Order_parent
         $this->afterOrderCleanUp($basket, $user);
 
         $isPayPalACDC = $paymentsId === PayPalDefinitions::ACDC_PAYPAL_PAYMENT_ID;
+        $isPaypalGooglePay = $paymentsId === PayPalDefinitions::GOOGLEPAY_PAYPAL_PAYMENT_ID;
         $isPayPalStandard = $paymentsId === PayPalDefinitions::STANDARD_PAYPAL_PAYMENT_ID;
         $isPaypalApplePay = $paymentsId === PayPalDefinitions::APPLEPAY_PAYPAL_PAYMENT_ID;
 
         $transactionId = null;
         $payPalPaymentSuccess = true;
 
-        if ($isPayPalACDC && $forceFetchDetails || $isPaypalApplePay) {
+        if ($isPayPalACDC && $forceFetchDetails || $isPaypalGooglePay  || $isPaypalApplePay) {
             $payPalApiOrder = $paymentService->fetchOrderFields($payPalOrderId);
             if ($this->isPayPalOrderCompleted($payPalApiOrder)) {
                 $this->markOrderPaid();
@@ -172,7 +173,7 @@ class Order extends Order_parent
             }
         }
 
-        if ($isPayPalACDC || $isPaypalApplePay) {
+        if ($isPayPalACDC || $isPaypalGooglePay || $isPaypalApplePay) {
             //webhook should kick in and handle order state and we should not call the api too often
             Registry::getSession()->deleteVariable(Constants::SESSION_ACDC_PAYPALORDER_STATUS);
             // remove PayPal order id from session
@@ -191,9 +192,7 @@ class Order extends Order_parent
                 $this->getId(),
                 $payPalOrderId,
                 $paymentsId,
-                PayPalApiOrder::STATUS_APPROVED,
-                '',
-                Constants::PAYPAL_TRANSACTION_TYPE_CAPTURE
+                PayPalApiOrder::STATUS_APPROVED
             );
         } else {
             // uAPM, PayPal Standard directly, PayPal Paylater
@@ -276,6 +275,9 @@ class Order extends Order_parent
         $isPayPalACDC = $sessionPaymentId === PayPalDefinitions::ACDC_PAYPAL_PAYMENT_ID;
         $isPayPalStandard = $sessionPaymentId === PayPalDefinitions::STANDARD_PAYPAL_PAYMENT_ID;
         $isPayPalPayLater = $sessionPaymentId === PayPalDefinitions::PAYLATER_PAYPAL_PAYMENT_ID;
+        if ($sessionPaymentId === PayPalDefinitions::GOOGLEPAY_PAYPAL_PAYMENT_ID) {
+            $isPayPalUAPM = false;
+        }
         if ($sessionPaymentId === PayPalDefinitions::APPLEPAY_PAYPAL_PAYMENT_ID) {
             $isPayPalUAPM = false;
         }